# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/code-sending": {
            "post": {
                "tags": [
                    "Forgot-password"
                ],
                "summary": "User Send Email",
                "description": "User Send Email",
                "operationId": "User Send Email",
                "requestBody": {
                    "description": "User Send Email",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "required"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "test@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/restore-password": {
            "post": {
                "tags": [
                    "Forgot-password"
                ],
                "summary": "restore-password",
                "description": "restore-password",
                "operationId": "restore-password",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "required"
                                ],
                                "properties": {
                                    "random_int": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "551485"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/update-password": {
            "post": {
                "tags": [
                    "Forgot-password"
                ],
                "summary": "update-password",
                "description": "update-password",
                "operationId": "update-password",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "required"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "3"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "password",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}